Implemented helper to create scenario from tensors.
Added ability to specify transforms in filelist scenario creation helper.

Fixed a bug in NCBatchInfo that prevented classes_in_this_batch field to be populated.

Fixed a bug that caused scenarios to wrongly return a Step when an out-of-bounds step id was given.

Fixed a bug in load_all_dataset.