Fix bug in GDumbPlugin
The base implementation was quite inefficient: for each new item,
it performed an entire concatenation with the current dataset.
This uses lists instead, and only concatenates once.

It also uses the task labels from each sample in the dataset, rather
the task label from the current experience, as there might be more
than one task in the dataset.

Signed-off-by: Fabrice Normandin <fabrice.normandin@gmail.com>